Allow to override build date
in order to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

Note that we leave cmake's TIMESTAMP call
because it already supports overriding the build time this way.

The Makefile.am change is designed to work with both GNU date and BSD date.