Configure travis to use docker for CI
Currently, there is an ubuntu 14.04 and archlinux build and they both
use cmake with ninja, cmake with make, and autotools to build. It should
be straightforward to add another configuration.

Travis checks out the correct branch of the source code for us, and we
expose that source code to the docker container using a docker run
"volume".