Add ability to draw a text cursor for the entry
Most of the call backs and procedures rely on the sheet being the one
in focus so prevent the entry being in focus. In doing this there is no
text cursor so add a draw function to draw one after the entry drawn.
Also remove the frame from the entry.