Fix problem with scrollbar sliders not tracking value
There seems to be a problem with scrollbar sliders do not track the
value and also if you try to drag the slider it does not move. This may
be related to bug https://bugzilla.gnome.org/show_bug.cgi?id=765410

What I have found is that this problem can be overcome by hiding and
showing the scrollbar so this commit does so.