report: render piechart with requested colors
A report can specify the pie chart colors with a call to:
   (gnc:html-piechart-set-colors! chart (list "blue" "red"))
However, these colors were ignored when rendering the piechart.

This patch adds the jqplot option "seriesColors" to pass
the requested colors to jqplot. If no colors are requested,
we let jqplot use default colors by setting "seriesColors: false".