report: render linechart with requested line colors
A report can specify the bar colors of a barchart with a call to:
   (gnc:html-linechart-set-col-colors! chart (list "blue" "red"))
However, these colors were ignored when rendering the linechart.

This patch adds the jqplot option "seriesColors" to pass
the requested bar colors to jqplot. If no colors are requested,
we let jqplot use default colors by setting "seriesColors: false".