report/net-linechart: calculate using gnc-monetary instead of double
Perform and store all calculations using gnc-monetary instead of double.
Conversion to double is only needed as a last step when adding data
to the chart (using new function "monetary->double").

If the table is displayed, its values are properly formatted as monetary
values since they are gnc-monetary instead of doubles.