use correct pointer (gnc_numeric**) with qof_instance_get
Function "qof_instance_get" needs a gnc_numeric** instead of a
gnc_numeric* to work correctly.
The proper usage of function "qof_instance_get" with a gnc_numeric
is shown in function:
  split-register-model.c:gnc_template_register_get_debcred_entry

This bug had an interesting side effect. It was causing the session
to be marked as dirty on startup: the Save button is active and an
asterisk (*) appears in the window title.
One of the last steps when reading the user data file is to replay
any scheduled transaction. This bug produced wrong values for
"sx-debit-numeric" and "sx-credit-numeric" which forced the transaction
to be modified marking the whole session as dirty.