cashflow-barchart: delete unnecessary intermediate lists
These intermediate lists *-value-list hold the double values needed
by the barchart. We can easily create these lists on the fly,
there's no need to store them in a separate variable.

The same is done to include total values in the table, we can easily
append the total value on the fly.