report-test: move add-to-load-path to build system
The unit test for cashflow-barchart.scm needs module stylesheets
to work, to render gnc-monetary objects correctly.
It makes sense to add the extra load-path needed by the module
in the build system instead of with code in the test file.

This is only needed when building with autotools. It seems cmake
builds don't need this.