report: fix default colors incorrectly set to transparent
The default color of a color option is set during a call to "gnc:make-color-option"
as a list of rgba values. In all color option instances in the scheme code this
default color is incorrectly set to transparent.
For example, white color was being set with:
  (list #xff #xff #xff 0)
The last value in the list is the value of the alpha channel, with 0 meaning
transparent. When the color button is displayed in the gui it shows a checkerboard
pattern instead of the intended color.

This commit changes all default colors of color options to have alpha=#xff, i.e.,
fully opaque.