Bug 771667 - Change reconciled splits warning
This patch displays two distinct warnings when changing protected
fields of a transaction that contains reconciled splits. If the fields
date, num and description are changed, then the warning list the
accounts that have reconciled splits and also advises that they will be
unreconciled after editing the transaction. If the fields account,
transfer, debit or credit are changed then the warning advises that the
split will be unreconciled after editing the transaction.

There is still just one warning preference as it is all to do with
fields protected by reconciliation.