Rework source directory structure for gschema, gtkbuilder and ui files
There is now a directory for each respective file type directly in the gnucash source directory
(This was already started for gschema files before).

In principle all files of each type should go into these global directories. The only
exception is if such a file belongs to an optional module. In that case the file
remains located inside that module's directory. Currently we have ui, glade and gschema file
exceptions in aqb and ofx importers.