Add spacing between border and text in dialog "New Accounts Hierarchy Setup".
There is no spacing between the text and the frame in the dialog
"New File" > "New Accounts Hierarchy Setup". In some cases - when
changing the window size - the text appears next to the frame.
It does not look very clean. As a solution, I have defined here a small spacing: 5px.