Bug 794360 Critical error when deleting accounts with children
When an account is deleted which has sub accounts a dialog is displayed
that had parts destroyed so they are not shown depending on certain
criteria and as a result an error occurred as it tried to get the
sensitivity of a destroyed widget. Change this so all parts of the
dialog are present but made in-sensitive when not applicable so the
dialog layout is the same on all occurrences.