TR: Price Column shown in commodity SCU
xaccSplitGetSharePrice will typically return a gnc-numeric (or scm
rational) value. This leads to the TR showing a price of, eg. $22 +
12228521/25349850 which is not ideal. This commit will modify the
share price to round the exact rational to an inexact decimal
according to the account share price currency SCU. If the SCU is less
than 10000, increase the precision by 2 significant digits.