string-sanitize: (gnc:render-options-changed) -> report-utilites.scm
We want to sanitize render-options-changed, therefore it must return
an html-object. Unfortunately this is not accessible to
app-utils/options.scm. If we move this function to
report-system/html-utilities.scm, it can access html-objects. Perhaps
it can be called with the make-generic-warning function in the future
to be displayed with the warning.