Bug 795519 - Credit card payment after reconciliation
The transfer dialog was set up as a child of the reconciliation dialog.
However as the latter is closed it's child dialogs are closed as well.
At least that's the behaviour on Windows. This didn't happen on Fedora 27.
Regardless it's fixed by making the account register the parent of the transfer
dialog.