[TR] [bugfix] disable grid if primary-key doesn't have subtotal
Formerly the logic would allow grid if secondary-subtotal enabled but
primary-subtotal disabled. This would produce a single-row grid. This
is now known to be inaccurate.

This is because there will be multiple secondary subtotals spread out
in the main table but no primary subtotal grouping to separate
them. This would make every subsequent secondary subtotal overwrite
the previous one. I think it is best to disable grid altogether if
primary-subtotal is not enabled.

The converse i.e. primary-subtotal enabled but secondary-subtotal
disabled, is handled correctly. It produces a single-column grid.