[TR] [bugfix] upgrade add-subtotal-row & grid for >1 commodities.
This aims to handle multiple commodities correctly. If a row contains
subtotals with more than one commodity, the row is duplicated so that
every commodity gets its own line in the grid.

This is accompanied by an upgrade to the test suite.

Sample output if prime-sortkey = accounts, sec-sortkey = monthly dates

        Jan-17  Feb-17  Mar-17   Total
Food    $22.00  $23.00  $35.00  $80.00
Books    $8.50   $9.55  $15.00  $33.05
Apps     £2.55   £5.00   £9.60  £17.15
Total                          $113.05
                                £17.15

(note monthly totals are not displayed because they're not actually
generated with the above sorting options)