Bug 795471 - Budget editor resize problems
In the original configuration, if you start with a wide account column
and drag it smaller and then make entry it pings back. The reason for
this is that the space is added to the total column to compensate as
the table width can not shrink. This can be fixed by removing the
viewport and stacking two scroll windows in a vbox and linking the
horizontal scrollbars.