Bug 796614 - Reconciliation report contains incorrect transactions
This commit modifies the date filter for reconciliation report,
ensuring only splits whose reconciled dates match report options.

If a split is not yet reconciled, include it anyway.

In a large datafile with a narrow reconciled date range, the
datefilter is likely to be the filter with highest frequency of #f
therefore it should be prioritised in the combined filter.

I chose to add another optional keyword instead of reusing the
existing #:custom-split-filter, because we need to detect when to
override the QofQuery date filter.