[TR] apply custom-sort after filtering.
As a follow on to last commit, if a large datafile is subject to a
reconciled date filter, the initial QofQuery date matcher would be
skipped, causing a large number of splits sent for custom
sorting prior to filtering.

It will always be more efficient that filtering is applied
first. Therefore custom sorting should be applied after filtering.