Bug 796665 - Backspace Key Inoperable After Ctrl+V
Remove the gtk_editable_set_editable toggles in
gnc_item_edit_cut_keyboard and gnc_item_edit_paste_keyboard. These were
added as a work-around to gnucash_sheet_key_press_event_internal leaving
the sheet not-editable after a regular key event. That was removed to
fix input-method key handling.

The original addition of that enable/disable toggle was apparently a
hack to fix the date-field accelerators and the use of the account
separator in the transfer field. Both seem to work fine with the
protection removed.