[balsheet-pnl] change strategy for col-data
Previously we were sending list-of-headers, and the length was used to
determine number of columns. The col-idx was sent to various callback
functions to retrieve column data.

Now we send the col-data - currently time64 dates or interval
datepairs. The num-columns is still (length col-data). The headers are
obtained via callback functions. All other callback functions are
amended to use col-data elements.