[balsheet-pnl] *needs testing - handle missing prices
Formerly, price conversion to a commodity would result in a
gnc-monetary with zero amount. This logic will, for
get-cell-amount-fn, attempt to convert an amount. If conversion fails
due to missing price, get-cell-amount-fn will return #f to indicate
missing cell amount.

The multilevel subtotals are also annotated with "(incomplete)" to
indicate some missing data.