Bug 771667 - Different warnings when changing reconciled splits vs. splits linked to reconciled splits
This is the final bit that ensures a warning is only presented if a previous warning
didn't already include the change we're about to warn about.

This should prevent two warnings popping up when only one is sufficient. For example
when the user first edits a protection field the warning will notify the user all
reconciled splits will be unreconciled. If the user then continues to edit a
(previously) a protected split field before leaving the transaction, no extra warning
is needed.