Add logo and favicon for gnucash bugzilla site
The logo is a blend of the gnucash tango icon with bugzilla's mascotte Buggie.
The favicon and assorted files has been generated from the logo via
https://realfavicongenerator.net
That site offers a more complete solution than just a simple favicon, when time
permits we may want to implement a more complete favicon handling on our websites
based on recommendations found there.