Increase of MAX_TAB_COUNT to change appearance of dialog "Book Options" in german language.
The dialog "Book Options" looks in german different, because it has 5 tabs: four standard tabs and additionaly one tab for tax settings, which appears only for german language. The 1 to 4 tabs appearance on top (horizontal mode). The appearance switchs to vertical mode (with tabs on the left side) for 5 and more tabs. Here the setting MAX_TAB_COUNT is increased to 5. This change should make the dialog similar in german and other languages.